requirements.txt
asgiref
django
django_bootstrap5
pip
djangorestframework
sqlparse
tzdata
python-decouple
psycopg2
aiogram
celery
requests